+++ src/test/scala/com/xorbit/spark/excel/ExcelReaderTest.scala
@@ -1,26 +1,10 @@
 package com.xorbit.spark.excel
 
-import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.types._
-import org.scalatest.BeforeAndAfterEach
-import org.scalatest.funsuite.AnyFunSuite
-import org.scalatest.matchers.should.Matchers
 
-class ExcelReaderTest extends AnyFunSuite with Matchers with BeforeAndAfterEach {
+class ExcelReaderTest extends SparkSessionLocal {
 
-  var spark : SparkSession = _
-  override def beforeEach() {
-    spark = SparkSession.builder().appName("udf testings")
-      .master("local[*]")
-      .config("", "")
-      .getOrCreate()
-  }
-
-  override def afterEach() {
-    spark.stop()
-  }
-
-  ignore("Read excel") {
+  test("Read excel") {
     spark.sparkContext.setLogLevel("ERROR")
 
     val schema = StructType(List(
@@ -58,7 +42,7 @@ class ExcelReaderTest extends AnyFunSuite with Matchers with BeforeAndAfterEach 
     }
   }
 
-  ignore("Read Sample.xlsx") {
+  test("Read Sample.xlsx") {
     spark.sparkContext.setLogLevel("ERROR")
     val schema = StructType(List(
       StructField("Id", IntegerType, true),
