+++ src/test/scala/com/xorbit/spark/excel/SparkSessionLocal.scala
@@ -1,16 +1,20 @@
 package com.xorbit.spark.excel
 
 import org.apache.spark.sql.SparkSession
+import org.scalatest.BeforeAndAfterEach
+import org.scalatest.funsuite.AnyFunSuite
+import org.scalatest.matchers.should.Matchers
 
-trait SparkSessionLocal {
-  def withSparkContext(testMethod: (SparkSession) => Any) {
-    val spark = SparkSession.builder()
+trait SparkSessionLocal extends AnyFunSuite with Matchers  with BeforeAndAfterEach {
+  var spark : SparkSession = _
+  override def beforeEach() {
+    spark = SparkSession.builder().appName("udf testings")
       .master("local[*]")
-      .appName("Spark test")
+      .config("", "")
       .getOrCreate()
-    try {
-      testMethod(spark)
-    }
-    finally spark.close()
+  }
+
+  override def afterEach() {
+    spark.stop()
   }
 }
